requirements.txt
python-telegram-bot==20.7
requests==2.31.1
flask==2.3.3
python-dotenv==1.0.0
phonenumbers==8.13.25